election-data-pipeline/01_upload_to_gcs/requirements.txt
google-cloud-storage==2.16.0
pandas==2.2.2
requests==2.31.0
